Introduced a moduke type scripting in the javascript files
