+++ src/modules/selectors.js
@@ -1,0 +1,2 @@
+export const listPlaceholder = document.querySelector('.list__placeholder');
+
